Rough crud machine working
Really gross/rough CRUD machine that works created.  Right now all that
is left to hit MVP would be cleaning up forms so that everything hides
and clears when the user is done.  I also need to clean up the backend
data model so it has a number of guests list.